+++ packman/apps/__init__.py
@@ -1,0 +1,1 @@
+from .predict_hinge import predict_hinge
